Revert "Properly name each artifact in GH Actions."
This reverts commit f293d1cd68f0b3457dbf3f4f793e9f9cadd5efa1.

An "artifact" in GitHub Actions is a collection of files, not a single
file, so the original version was working fine (with all files under the
collection named "artifact") whereas the new version broke because it
tried to download the non-existent artifact named "artifact".

Let's just revert to the old version because it was fine.